cmd/trayscale: refactor and get status switch working again
